[Feature #163966060] Check token blacklist auth during decode
